VersionsManager versions argument
VersionManager was not doing any value checking for it's *versions* keyword argument. If no versions were passed or an non-iterable was passed it would raise an Exception on VersionManager.register_versions(*versions).

Updated to change None/no argument to list() and a try/except to shallow copy versions, raising an error if it can't to properly explain that versions should be an iterable.